feat: removeQuotes
